analysis: add analysis to calculate reputation deltas of target w/ peers
For some setup:
---\
---> P1 ----+
---/        |
            T
---\        |
---> P2 ----+
---/

This program will find all the reputation pairs for each peer of target
T and calculate the reputation delta between the outgoing channel (with
the target)'s outgoing reputation and the incoming revenue of the peer's
incoming channel. This delta describes the total htlcs that the peer can
have in-flight to the target on its channels before they stop forwarding
payments to the target node (which we assume is being jammed
downstream).