analysis: get projected revenue from actual simulation run
analysis: pull projection from real data

data: add projections from real network

For the attacks where we add the attacker to the graph, we just copy
projections from the network where they aren't present. This is because
any traffic drawn in by an attacker is a side effect of the attack,
so we compare to the honest case where no attack has occurred.

Note: ln_50/ln_51 are generated with simln:0.2.0